refactor: split state-API client and frame rendering out of cmd/tui/main.go (#661)
Pure behavior-preserving move along the boundaries the #661
prioritization comment names for the TUI: the /api/v1/state client
(response DTOs, authenticated fetch, TPS sampling) moves to
state_client.go and the frame rendering (renderFrame plus the
table/format/terminal-width helpers) moves to render.go, leaving
main.go at 252 lines holding the poll loop and terminal lifecycle.
This pays down the LAST entry of the oversized-file baseline: the map
in scripts/file_size_budget_test.go is now empty and every production
Go file is enforced by the generic 800-line gate directly.
